% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{comparator}
\alias{comparator}
\title{Semantic version comparator}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x1061e8358> 
 - attr(*, "name")= chr "comparator_generator"
}}
\usage{
comparator
}
\description{
TODO
}
\keyword{datasets}

