% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compare}
\alias{!=.semver}
\alias{<.semver}
\alias{<=.semver}
\alias{==.semver}
\alias{>.semver}
\alias{>=.semver}
\alias{cmp}
\alias{compare}
\alias{compare_loose}
\alias{eq}
\alias{gt}
\alias{gte}
\alias{lt}
\alias{lte}
\alias{neq}
\alias{rev_compare}
\title{Compare semantic version numbers}
\usage{
compare(a, b, loose = FALSE)

compare_loose(a, b)

rev_compare(a, b, loose = FALSE)

gt(a, b, loose = FALSE)

lt(a, b, loose = FALSE)

eq(a, b, loose = FALSE)

neq(a, b, loose = FALSE)

gte(a, b, loose = FALSE)

lte(a, b, loose = FALSE)

cmp(a, op = "==", b, loose = FALSE)

\method{"<"}{semver}(lhs, rhs)

\method{">"}{semver}(lhs, rhs)

\method{"<="}{semver}(lhs, rhs)

\method{">="}{semver}(lhs, rhs)

\method{"=="}{semver}(lhs, rhs)

\method{"!="}{semver}(lhs, rhs)
}
\arguments{
\item{a}{Version string or \code{semver} object.}

\item{b}{Version string or \code{semver} object.}

\item{loose}{Whether loose ranges are allowed.}

\item{op}{Operator to test for. Possible values:
\code{=} and \code{==} are equivalent, the empty string also means
\code{==}; \code{!=}, \code{>}, \code{>=}, \code{<}, \code{<=}.}
}
\value{
\code{-1} if \code{a} is older, \code{1} if \code{a} is
  newer, \code{0} if \code{a} and \code{b} are equal.

Logical scalar.
}
\description{
TODO
}

