% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inc}
\alias{inc}
\title{Increase a semantic version number}
\usage{
inc(version, release, loose = FALSE)
}
\arguments{
\item{version}{Version number string or \code{semver} object. The
version number to increase.}

\item{release}{Which component to increase. See \code{\link{semver}} and
the \code{inc} method for possible values.}

\item{loose}{Whether loose ranges are allowed.}
}
\description{
TODO
}

