% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ltr}
\alias{ltr}
\title{Check if a version is older than a range requirement}
\usage{
ltr(version, range, loose = FALSE)
}
\arguments{
\item{version}{Version string or \code{semver} object.}

\item{range}{Range requirement string or \code{range} object.}

\item{loose}{Whether loose ranges are allowed.}
}
\description{
TODO
}

