% Generated by roxygen2 (4.0.1): do not edit by hand
\name{max_satisfying}
\alias{max_satisfying}
\title{Choose the newest version that satisfies a range requirement}
\usage{
max_satisfying(versions, range, loose = FALSE)
}
\arguments{
\item{versions}{List or vector of version strings or \code{semver}
objects.}

\item{range}{Range requirement string or \code{range} object.}

\item{loose}{Whether non-strict versions are allowed.}
}
\description{
TODO
}

