% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{range}
\alias{range}
\title{Range specification}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x1060d9408> 
 - attr(*, "name")= chr "range_generator"
}}
\usage{
range
}
\description{
TODO
}
\keyword{datasets}

