% Generated by roxygen2 (4.0.1): do not edit by hand
\name{satisfies}
\alias{satisfies}
\title{Check if a semantic version number satisfies a range requirement}
\usage{
satisfies(version, range, loose = FALSE)
}
\arguments{
\item{version}{Version string or \code{semver} object.}

\item{range}{Range requirement string or \code{range} object.}

\item{loose}{Whether non-strict versions are allowed.}
}
\description{
TODO
}

