% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{semver-package}
\alias{semver-package}
\title{Semantic versioning for R packages}
\description{
Semantic versioning is a method for consistently assigning
version numbers to various builds of your software. See
\url{http://semver.org} for details.
}
\details{
This package includes utilities to deal with semantic versioning
tasks: \itemize{
  \item Validate version numbers, \code{\link{valid}}.
  \item Decide if a version is newer than another one,
    \code{\link{gt}}, \code{\link{lt}}.
  \item Increasing a version number, \code{\link{inc}}.
  \item Sorting a list of version numbers, \code{\link{semver_sort}}.
  \item Tools to specifies ranges of version numbers,
    \code{\link{range}}.
  \item Tools to check if a version number satisfied a range
    requirement, \code{\link{satisfies}}.
}

The API and the implementation closely follows the
Javascript semantic versioning implementation for the npm
package manager, see \url{https://github.com/npm/node-semver}.
}

