% Generated by roxygen2 (4.0.1): do not edit by hand
\name{[.semver_list}
\alias{==.semver_list}
\alias{>.semver_list}
\alias{[.semver_list}
\alias{semver_sort}
\title{Sort semantic version numbers}
\usage{
\method{"["}{semver_list}(x, i)

\method{">"}{semver_list}(a, b)

\method{"=="}{semver_list}(a, b)

semver_sort(list, loose = FALSE)
}
\arguments{
\item{list}{List or vector of version number strings or \code{semver}
objects. You can also mix them.}

\item{loose}{Whether loose ranges are allowed.}
}
\value{
Sorted list or vector of versions numbers or \code{semver}
  objects.
}
\description{
TODO
}

