% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{SEMVER_SPEC_VERSION}
\alias{SEMVER_SPEC_VERSION}
\title{The version of the semantic versioning standard this package implements}
\format{\preformatted{ chr "2.0.0"
}}
\usage{
SEMVER_SPEC_VERSION
}
\description{
See \url{http://semver.org} for details.
}
\keyword{datasets}

