% Generated by roxygen2 (4.0.1): do not edit by hand
\name{valid_range}
\alias{valid_range}
\title{Check if a range requirement specification is valid}
\usage{
valid_range(range, loose = FALSE)
}
\arguments{
\item{range}{Range requirement string.}

\item{loose}{Whether loose ranges are allowed.}
}
\description{
TODO
}

