% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semver-package.r
\docType{package}
\name{semver}
\alias{semver}
\alias{semver-package}
\title{\pkg{semver} parse semantic versions.}
\description{
\code{semver} has two functions:
\code{\link{parse_version}},
\code{\link{render_version}},
}
\details{
\code{\link{parse_version}} takes as input a character vector with
elements following the specification at \url{http://semver.org/}.

Comparisons can be done on the "svlist" and "svptr" classes returned
by \code{\link{parse_version}}. \code{\link{Ops}} and
\code{\link{summary}} methods are implemented for both classes.
}

